[feat-1]Add A Burger And Lasagna Recipe Page
Create 2 different html files. One containing the recipe for a burger
which is formatted with h1, h2, p, ul, ol, li, etc. tags. The other
containing the recipe for lasagna and is formatted in the same way.
These are both then linked using the <a> tag to index.html or the
home page of the site.

each recipe contains a photo, a list of ingredients, and the steps
to making each dish.